pyproject.toml
[tool.poetry]
name = "imk"
version = "0.1.0"
description = ""
authors = ["Tom Zheng <tomzheng77@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
jupyter = "^1.0.0"
llama-index = "^0.10.37"
langchain = "^0.1.20"
openai = "^1.30.1"
python-dotenv = "^1.0.1"
trafilatura = "^1.9.0"
iteration-utilities = "^0.12.1"
forbiddenfruit = "^0.1.4"
langchain-openai = "^0.1.7"
langchain-chroma = "^0.1.1"
llama-index-llms-langchain = "^0.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

